+++ src/atoms/generateList.atom.ts
@@ -1,0 +1,7 @@
+import { atom } from 'recoil';
+import { GenerateStudentsListType } from 'types/user.type';
+
+export const generateListState = atom<GenerateStudentsListType[]>({
+  key: 'generateListState',
+  default: [],
+});
