+++ src/features/home/generateStudents.feature.ts
@@ -1,8 +1,10 @@
 import { generateStudnets } from 'apis/user.api';
+import { generateListState } from 'atoms/generateList.atom';
 import { useModal } from 'hooks/useModal';
 import { useErrorToast } from 'hooks/useToast';
 import { Dispatch, SetStateAction } from 'react';
 import { useMutation } from 'react-query';
+import { useSetRecoilState } from 'recoil';
 import { GenerateStudnetsType } from 'types/user.type';
 
 interface GenerateStudentsFeaturePropsType {
@@ -14,12 +16,14 @@ export const GenerateStudentsFeature = ({
   generateStudentsData,
   setIsOpenCheckGenerateModal,
 }: GenerateStudentsFeaturePropsType) => {
+  const setGenerateListData = useSetRecoilState(generateListState);
   const openCheckGenerateModal = () => setIsOpenCheckGenerateModal(true);
   const { closeModal } = useModal();
 
   const { mutate } = useMutation(generateStudnets, {
     onSuccess: (res) => {
       console.log(res);
+      setGenerateListData(res);
       openCheckGenerateModal();
     },
     onError: (err) => {
